+++ hybrid_vision/hybrid_vision/example/lib/main.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'view/view.dart';
 
 void main() {
-  WidgetsFlutterBinding.ensureInitialized();
   runApp(const MyApp());
 }