+++ hybrid_vision/hybrid_vision/lib/hybrid_vision.dart
@@ -1,2 +1,3 @@
-export 'package:hybrid_vision_platform_interface/hybrid_vision_platform_interface.dart'
-    hide BarcodePlatform;
+export 'package:hybrid_vision_platform_interface/hybrid_vision_platform_interface.dart';
+export 'package:hybrid_vision_android/hybrid_vision_android.dart';
+export 'package:hybrid_vision_darwin/hybrid_vision_darwin.dart';