+++ hybrid_vision/hybrid_vision_android/lib/src/_init.dart
@@ -1,8 +1,0 @@
-// ignore_for_file: use_super_parameters
-import "dart:ffi" as ffi;
-import "package:jni/internal_helpers_for_jnigen.dart";
-
-// Auto-generated initialization code.
-
-final ffi.Pointer<T> Function<T extends ffi.NativeType>(String sym) jniLookup =
-    ProtectedJniExtensions.initGeneratedLibrary("hybrid_vision_android");