+++ hybrid_vision/hybrid_vision_android/lib/src/android/graphics/_package.dart
@@ -1,4 +1,0 @@
-export "Bitmap.dart";
-export "BitmapFactory.dart";
-export "Point.dart";
-export "Rect.dart";