+++ hybrid_vision/hybrid_vision_android/lib/src/com/google/mlkit/vision/barcode/_package.dart
@@ -1,4 +1,0 @@
-export "BarcodeScanner.dart";
-export "BarcodeScannerOptions.dart";
-export "BarcodeScanning.dart";
-export "ZoomSuggestionOptions.dart";