+++ hybrid_vision/hybrid_vision_darwin/lib/hybrid_vision_darwin.dart
@@ -1,9 +1,2 @@
-import 'package:hybrid_vision_platform_interface/hybrid_vision_platform_interface.dart';
-
-import 'src/barcode.dart';
-
-abstract class HybridVisionDarwin {
-  static void registerWith() {
-    BarcodePlatform.instance = BarcodePlatformImpl();
-  }
-}
+export 'src/hybrid_vision_darwin_plugin.dart';
+export 'src/barcode_detection.dart';