+++ hybrid_vision/hybrid_vision/CHANGELOG.md
@@ -1,6 +1,6 @@
-## 0.0.1
+## 0.1.0
 
-* Add `BarcodeDetector` class.
+* `Android` Fix the crash issue caused by r8 on release mode.
 
 ## 0.0.2
 
@@ -8,3 +8,7 @@
 * `Android` Fix the crash issue caused by `BarcodeDetector#detect`.
 * `iOS` Fix the example crash issue caused by missed the `NSPhotoLibraryUsageDescription`.
 * `macOS` Update `hybrid_core` to 0.1.0.
+
+## 0.0.1
+
+* Add `BarcodeDetector` class.